django/ToDo-List-Django-REST-Zhoomart/ToDo-List-Django-REST-main/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
openapi==1.1.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
